refactor: move root store to /models
